move shared utility to shared space
