+++ custom/ilsgateway/utils.py
@@ -73,12 +73,3 @@ def send_translated_message(user, message, **kwargs):
     with localize(user.get_language_code()):
         send_sms_to_verified_number(verified_number, message % kwargs)
         return True
-
-
-def get_supply_point_by_external_id(domain, external_id):
-    return SupplyPointCase.view('hqcase/by_domain_external_id',
-        key=[domain, str(external_id)],
-        reduce=False,
-        include_docs=True,
-        limit=1
-    ).first()
